separate spectral.py -> bandpasses.py, magsystems.py, spectrum.py
